Putting everything except the '_transform_pw_linear_expr' method in a base class that all the PW to GDP transformations can use.
