+++ pyomo/contrib/piecewise/transform/inner_representation_gdp.py
@@ -9,30 +9,21 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
-from pyomo.common.collections import ComponentMap
 from pyomo.common.config import ConfigDict, ConfigValue
-from pyomo.common.modeling import unique_component_name
 from pyomo.contrib.fbbt.fbbt import compute_bounds_on_expr
-from pyomo.contrib.piecewise import PiecewiseLinearFunction
-from pyomo.contrib.piecewise.transform.piecewise_to_mip_visitor import (
-    PiecewiseLinearToMIP)
-from pyomo.core import (
-    Constraint, Objective, Var, BooleanVar, Expression, Suffix, Param, Set,
-    SetOf, RangeSet, ExternalFunction, Connector, SortComponents, Any,
-    NonNegativeIntegers, NonNegativeReals)
-from pyomo.core.base import Transformation, TransformationFactory
-from pyomo.core.base.block import _BlockData, Block
+from pyomo.contrib.piecewise.transform.piecewise_to_gdp_transformation import (
+    PiecewiseLinearToGDP)
+from pyomo.core import Constraint, NonNegativeIntegers, Suffix, Var
+from pyomo.core.base import TransformationFactory
 from pyomo.core.util import target_list
 from pyomo.gdp import Disjunct, Disjunction
-from pyomo.gdp.util import is_child_of
-from pyomo.network import Port
 
 @TransformationFactory.register('contrib.inner_repn_gdp',
                                 doc="Convert piecewise-linear model to a GDP "
                                 "using an inner representation of the "
                                 "simplices that are the domains of the linear "
                                 "functions.")
-class InnerRepresentationGDPTransformation(Transformation):
+class InnerRepresentationGDPTransformation(PiecewiseLinearToGDP):
     """
     Convert a model involving piecewise linear expressions into a GDP by
     representing the piecewise linear functions as Disjunctions where the
@@ -85,115 +76,7 @@ class InnerRepresentationGDPTransformation(Transformation):
         transformed.
         """
     ))
-    def __init__(self):
-        super().__init__()
-        self.handlers = {
-            Constraint: self._transform_constraint,
-            Objective: self._transform_objective,
-            Var:         False,
-            BooleanVar:  False,
-            Connector:   False,
-            Expression:  False,
-            Suffix:      False,
-            Param:       False,
-            Set:         False,
-            SetOf:       False,
-            RangeSet:    False,
-            Disjunction: False,
-            Disjunct:    False,
-            Block:       self._transform_block,
-            ExternalFunction: False,
-            Port:        False,
-            PiecewiseLinearFunction: self._transform_piecewise_linear_function,
-        }
-        self._transformation_blocks = {}
-
-    def _apply_to(self, instance, **kwds):
-        try:
-            self._apply_to_impl(instance, **kwds)
-        finally:
-            self._transformation_blocks.clear()
-
-    def _apply_to_impl(self, instance, **kwds):
-        config = self.CONFIG(kwds.pop('options', {}))
-        config.set_value(kwds)
-
-        targets = config.targets
-        if targets is None:
-            targets = (instance, )
-
-        knownBlocks = {}
-        not_walking_exprs_msg = (
-            "When not descending into expressions, Constraints "
-            "and Objectives are not valid targets. Please specify "
-            "PiecewiseLinearFunction component and the Blocks "
-            "containing them, or (at the cost of some performance "
-            "in this transformation), set the 'descend_into_expressions' "
-            "option to 'True'.")
-        for t in targets:
-            if not is_child_of(parent=instance, child=t,
-                               knownBlocks=knownBlocks):
-                raise ValueError("Target '%s' is not a component on instance "
-                                 "'%s'!" % (t.name, instance.name))
-            if t.ctype is PiecewiseLinearFunction:
-                if config.descend_into_expressions:
-                    raise ValueError(
-                        "When descending into expressions, the transformation "
-                        "cannot take PiecewiseLinearFunction components as "
-                        "targets. Please instead specify the Blocks, "
-                        "Constraints, and Objectives where your "
-                        "PiecewiseLinearFunctions have been used in "
-                        "expressions.")
-                self._transform_piecewise_linear_function(
-                    t, config.descend_into_expressions)
-            elif t.ctype is Block or isinstance(t, _BlockData):
-                self._transform_block(t, config.descend_into_expressions)
-            elif t.ctype is Constraint:
-                if not config.descend_into_expressions:
-                    raise ValueError(
-                        "Encountered Constraint target '%s':\n%s"
-                        % (t.name, not_walking_exprs_msg))
-                self._transform_constraint(t, config.descend_into_expressions)
-            elif t.ctype is Objective:
-                if not config.descend_into_expressions:
-                    raise ValueError(
-                        "Encountered Objective target '%s':\n%s"
-                        % (t.name, not_walking_exprs_msg))
-                self._transform_objective(t, config.descend_into_expressions)
-            else:
-                raise ValueError(
-                    "Target '%s' is not a PiecewiseLinearFunction, Block or "
-                    "Constraint. It was of type '%s' and can't be transformed."
-                    % (t.name, type(t)))
-
-    def _get_transformation_block(self, parent):
-        if parent in self._transformation_blocks:
-            return self._transformation_blocks[parent]
-
-        nm = unique_component_name(
-            parent,
-            '_pyomo_contrib_pw_linear_inner_repn')
-        self._transformation_blocks[parent] = transBlock = Block()
-        parent.add_component(nm, transBlock)
-
-        transBlock.transformed_functions = Block(Any)
-        return transBlock
-
-    def _transform_block(self, block, descend_into_expressions):
-        blocks = block.values() if block.is_indexed() else (block,)
-        for b in blocks:
-            for obj in b.component_objects(
-                    active=True,
-                    descend_into=(Block, Disjunct),
-                    sort=SortComponents.deterministic):
-                handler = self.handlers.get(obj.ctype, None)
-                if not handler:
-                    if handler is None:
-                        raise RuntimeError(
-                            "No transformation handler registered for modeling "
-                            "components of type '%s'." % obj.ctype)
-                    continue
-                handler(obj, descend_into_expressions)
+    _transformation_name = 'pw_linear_inner_repn'
 
     def _transform_pw_linear_expr(self, pw_expr, pw_linear_func,
                                   transformation_block):
@@ -245,48 +128,3 @@ class InnerRepresentationGDPTransformation(Transformation):
             expr=[d for d in transBlock.disjuncts.values()])
 
         return transBlock.substitute_var
-
-    def _transform_piecewise_linear_function(self, pw_linear_func,
-                                             descend_into_expressions):
-        if descend_into_expressions:
-            return
-
-        transBlock = self._get_transformation_block(
-            pw_linear_func.parent_block())
-        _functions = pw_linear_func.values() if pw_linear_func.is_indexed() \
-                     else (pw_linear_func,)
-        for pw_func in _functions:
-            for pw_expr in pw_func._expressions.values():
-                substitute_var = self._transform_pw_linear_expr(pw_expr.expr,
-                                                                pw_func,
-                                                                transBlock)
-                # We change the named expression to point to the variable that
-                # will take the appropriate value of the piecewise linear
-                # function.
-                pw_expr.expr = substitute_var
-
-    def _transform_constraint(self, constraint, descend_into_expressions):
-        if not descend_into_expressions:
-            return
-
-        transBlock = self._get_transformation_block(constraint.parent_block())
-        visitor = PiecewiseLinearToMIP(self._transform_pw_linear_expr,
-                                       transBlock)
-
-        _constraints = constraint.values() if constraint.is_indexed() else \
-                       (constraint,)
-        for c in _constraints:
-            visitor.walk_expression((c.expr, c, 0))
-
-    def _transform_objective(self, objective, descend_into_expressions):
-        if not descend_into_expressions:
-            return
-
-        transBlock = self._get_transformation_block(objective.parent_block())
-        visitor = PiecewiseLinearToMIP(self._transform_pw_linear_expr,
-                                       transBlock)
-
-        _objectives = objective.values() if objective.is_indexed() else \
-                      (objective,)
-        for o in _objectives:
-            visitor.walk_expression((o.expr, o, 0))
